% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-wrap.R
\name{DBI-wrap}
\alias{DBI-wrap}
\alias{dbDataType,Lease-method}
\alias{dbGetQuery,Lease,ANY-method}
\alias{dbExecute,Lease,ANY-method}
\alias{dbListFields,Lease,ANY-method}
\alias{dbListTables,Lease-method}
\alias{dbListObjects,Lease-method}
\alias{dbReadTable,Lease,ANY-method}
\alias{dbWriteTable,Lease,ANY-method}
\alias{dbCreateTable,Lease-method}
\alias{dbAppendTable,Lease-method}
\alias{dbExistsTable,Lease,ANY-method}
\alias{dbRemoveTable,Lease,ANY-method}
\alias{dbIsReadOnly,Lease-method}
\alias{sqlData,Lease-method}
\alias{sqlCreateTable,Lease-method}
\alias{sqlAppendTable,Lease-method}
\alias{sqlInterpolate,Lease-method}
\alias{sqlParseVariables,Lease-method}
\alias{dbQuoteIdentifier,Lease,ANY-method}
\alias{dbUnquoteIdentifier,Lease-method}
\alias{dbQuoteLiteral,Lease-method}
\alias{dbQuoteString,Lease,ANY-method}
\alias{dbAppendTableArrow,Lease-method}
\alias{dbCreateTableArrow,Lease-method}
\alias{dbGetQueryArrow,Lease-method}
\alias{dbReadTableArrow,Lease-method}
\alias{dbSendQueryArrow,Lease-method}
\alias{dbWriteTableArrow,Lease-method}
\title{DBI methods (simple wrappers)}
\usage{
\S4method{dbDataType}{Lease}(dbObj, obj, ...)

\S4method{dbGetQuery}{Lease,ANY}(conn, statement, ...)

\S4method{dbExecute}{Lease,ANY}(conn, statement, ...)

\S4method{dbListFields}{Lease,ANY}(conn, name, ...)

\S4method{dbListTables}{Lease}(conn, ...)

\S4method{dbListObjects}{Lease}(conn, prefix = NULL, ...)

\S4method{dbReadTable}{Lease,ANY}(conn, name, ...)

\S4method{dbWriteTable}{Lease,ANY}(conn, name, value, ...)

\S4method{dbCreateTable}{Lease}(conn, name, fields, ..., row.names = NULL, temporary = FALSE)

\S4method{dbAppendTable}{Lease}(conn, name, value, ..., row.names = NULL)

\S4method{dbExistsTable}{Lease,ANY}(conn, name, ...)

\S4method{dbRemoveTable}{Lease,ANY}(conn, name, ...)

\S4method{dbIsReadOnly}{Lease}(dbObj, ...)

\S4method{sqlData}{Lease}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{Lease}(con, table, fields, row.names = NA, temporary = FALSE, ...)

\S4method{sqlAppendTable}{Lease}(con, table, values, row.names = NA, ...)

\S4method{sqlInterpolate}{Lease}(conn, sql, ..., .dots = list())

\S4method{sqlParseVariables}{Lease}(conn, sql, ...)

\S4method{dbQuoteIdentifier}{Lease,ANY}(conn, x, ...)

\S4method{dbUnquoteIdentifier}{Lease}(conn, x, ...)

\S4method{dbQuoteLiteral}{Lease}(conn, x, ...)

\S4method{dbQuoteString}{Lease,ANY}(conn, x, ...)

\S4method{dbAppendTableArrow}{Lease}(conn, name, value, ...)

\S4method{dbCreateTableArrow}{Lease}(conn, name, value, ..., temporary = FALSE)

\S4method{dbGetQueryArrow}{Lease}(conn, statement, ...)

\S4method{dbReadTableArrow}{Lease}(conn, name, ...)

\S4method{dbSendQueryArrow}{Lease}(conn, statement, ...)

\S4method{dbWriteTableArrow}{Lease}(conn, name, value, ...)
}
\arguments{
\item{dbObj}{A DBI Driver][DBI::DBIDriver-class] or
\link[DBI:DBIConnection-class]{DBI Connection}.}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Other arguments passed on to methods.}

\item{conn}{A \link[DBI:DBIConnection-class]{DBI Connection}.}

\item{statement}{a character string containing SQL.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{prefix}{A fully qualified path in the database's namespace, or \code{NULL}.
This argument will be processed with \code{\link[DBI:dbUnquoteIdentifier]{dbUnquoteIdentifier()}}.
If given the method will return all objects accessible through this prefix.}

\item{value}{A \link{data.frame} (or coercible to data.frame).}

\item{fields}{Either a character vector or a data frame.

A named character vector: Names are column names, values are types.
Names are escaped with \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}.
Field types are unescaped.

A data frame: field types are generated using
\code{\link[DBI:dbDataType]{dbDataType()}}.}

\item{row.names}{Must be \code{NULL}.}

\item{temporary}{If \code{TRUE}, will generate a temporary table.}

\item{con}{A database connection.}

\item{table}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{values}{A data frame. Factors will be converted to character vectors.
Character vectors will be escaped with \code{\link[DBI:dbQuoteString]{dbQuoteString()}}.}

\item{sql}{A SQL string containing variables to interpolate.
Variables must start with a question mark and can be any valid R
identifier, i.e. it must start with a letter or \code{.}, and be followed
by a letter, digit, \code{.} or \verb{_}.}

\item{.dots}{A list of named arguments to interpolate.}

\item{x}{A character vector, \link[DBI]{SQL} or \link[DBI]{Id} object to quote as identifier.}
}
\description{
DBI methods (simple wrappers)
}
