% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{dbLease}
\alias{dbLease}
\title{Create a lease to a database connection}
\usage{
dbLease(drv, ..., timeout = 60, max_retries = 10)
}
\arguments{
\item{drv}{A \link[DBI:DBIDriver-class]{DBI Driver}, e.g. \code{duckdb::duckdb()}}

\item{...}{Additional arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{timeout}{Seconds to hold the connection open for reuse.}

\item{max_retries}{Number of times to retry connecting using exponential back-off.}
}
\value{
An object of class lease
}
\description{
Create a lease to a database connection
}
